Changes to models.py:
added 2 classes, PostJob and Application.

The PostJob class takes in the company name, job title, job location,
employment type, seniority level, and job description from the
employer and formats it in the data base.

The Application class takes i nthe name, email, phone number, and
the short note to the company that the applicant has given and formats
it in the data base.